Fix version check for projects with multiple application types
Deploying older versions of an application is not allowed, so the deployed
versions are retrieved for all application types for a given project and
then compared.  The problem was if there were multiple application types
it was checking again ALL of them instead of just the ones being requested.
This patch fixes that and also ensures that if the application types weren't
specified on the command line (args.apptypes will be None), it does check
against all application types for the project.