Better handling of deployment to previous environment check
If a version of an application being deployed to stage or production
hasn't been deployed to the previous environment, TDS tracebacks with
an out of index error.  This fixes that, though it has a small issue
where if multiple apptypes are being deployed to, the info message
won't say which one (to be fixed in the refactor).