Rework repo daemon and TDS logic
The basic logic for update_deploy_repo and TDS's 'tds package add' has
been redone to improve stability and reliability.  Not all changes are
yet in place; the DB updates from the daemon itself aren't yet in place
as this will require refactoring the config file parsing and DB ini-
tialization to be usable from both TDS and the daemon, but the basic
pieces are now in place.