Update changes to update_deploy_repo daemon
Fix some issues with the update_deploy_repo daemon and an additional fix
into the 'add()' command in package to ensure the database is re-read
when checking for changes for package status.

Also added signal handlers for shutdown for when zookeeper is in use
(otherwise the daemon requires a SIGKILL).