Attempt to avoid most fatal errors
As per advice from Chris, a first attempt to avoid most of the fatal
errors that could cause the daemon to exit.  I have not been completely
successful, but believe I have removed most of them and reworked some
of the logic to potentially allow multiple servers to be processing
at the same time (though there still could be an issue if two 'make's
are run at the same time).