Correct use of 'targets' to 'target' for host objects
The hosts table now uses 'target' instead of 'targets' (a host can have
only one apptype), and the previous change for this missed quite a few
uses of the multiple type, which has been fixed.

Also added was a 'package_defintions' property on the HostTarget model,
needed for testing equality in the 'validate_targets' method.