Validate package downloaded from Jenkins
Sometimes Jenkins does not give a complete transfer of the RPM artifact,
leading to bad RPMs being introduced into the system.  In order to mitigate
this, up to three attempts to do the transfer are done with each being
checked via Jenkins fingerprinting mechanism.  Only if all three transfers
fail will 'tds package add' abort.

If the fingerprinting has not been enabled for a given application, there
is a warning message printed with a reference on how to do so, but does
allow the user to continue if desired; this will eventually be a fatal
error in the program once all Jenkins jobs have been adjusted accordingly.