Fix race condition with TDS and update repo daemon
TDS was doing a basic validation check on the RPM downloaded after it
moved it to the 'incoming' directory.  Unfortunately if the update repo
daemon sees the file and moves it to the 'processing' directory, TDS
will apparently fail due to the race condition even though the package
add does succeed.  Moving the check into the _queue_rpm() method and
checking the file after it's been written out to the 'tmp' directory
will prevent this.