Make the feature tests for 'tds deploy invalidate' work
The original tests actually displayed a bug in the logic regarding what
could be invalidated (the current deployment was having this done to it
with the new code) and the tests were insufficient, so they were expanded
to do the right thing.

One test is disabled until the 'deploy with no hosts in tier' decision
is made, and a model step was fixed that was actually invalid (though
not being used).  A few other minor adjustments made to ensure things
worked properly.