No, really, make the unvalidated program work
Apparently the new code was never properly tested before, so once other
issues were resolved, a new set of problems cropped up (namely in the
setup of the 'fake' deployment object).  These have been fixed, along
with a logic flaw in determining the latest unvalidated deployments for
a given tier and application, so this should now function correctly.