Add 10 second timeout for slack/signalfx notification calls
Motivation: Make the release process consistently fast for better
productivity. The old behavior was to have no timeout.

I assume the root cause might be some sort of issue with the proxy or
signalfx/slack APIs. I'm not sure why this is the only affected application.

And use octal to avoid a confusing python3 tokenizing error with mypy
(works in both python 2 and 3)
(could use `mypy --py2` if there were any plans to set that up)

ORA2PG-438