SRE-1154: rewrite process cleanup
bug fixes:
1. Do not commit session unless dep.status is changed. There is no need
to commit a lack of change. Note a related deficiency here, though.

2. Do not terminate processes after they have written status to the
database. Be nice and let them finish on their own.

3. Attempt to kill processes when they do not terminate.

4. Reap child processes rather than forgetting about them and leaving
zombies around. The subprocess module does some internal reaping upon
process creation, but that can take a long time to happen.

To make this work, rewrite the cleanup method into three phases all
contained with the main loop. This removes some duplicate
database-setting code, but is longer in general in order to implement
the better process management. The python 2 subprocess module does not
support non-blocking waits, so we have to do it the hard way.