SRE-15125: convert update_deploy_repo to TDSDaemon (with systemd)
Make update_deploy_repo be a child of TDSDaemon in order to catch it up
with improvements to functionality shared with tds_installer.

* run without internal daemonization
* fix zookeeper host specification
* enable zookeeper logging
* fix zookeeper connection handling
* use a zookeeper lock instead of election
* stop on SIGINT

See the history of tds/scripts/tds_installer.py for details.

The change to run without internal daemonization is paired with a
conversion to systemd (they are mutually dependent). The rest of the
systemd change is largely based on the coresponding change to
tds_installer, afdc90094.