SRE-18741: improve handling of zookeeper session state changes
Previously, we detected session loss in a callback and handled that by
releasing the lock. This is a good intention, but the callback in turn
called back into kazoo, which:
1. Had no valid session with which to remove the corresponding zookeeper
   node.
2. Is not configured to retry.
3. Therefore raised an exception.
4. ...which is caught by the overlying kazoo code.
5. ...so the rest of our callback never ran, and we kept self.has_lock
   == True.

Having the callback call back into the library which ran the callback is
probably not a good idea anyway.

Instead, have the callback merely drop our own notion of having a lock.
To be safe, remove any potentially-remaining locks before lock
acquisition. In order to prevent the cancellation part of this from
cancelling the _subsequent_ lock acquisition, break out the cancellation
into its own method as well.