SRE-21810: ensure that lock is released
If there's zookeeper connection trouble, the daemon code figured the
lock was gone, but apparently, that is not always true. When a lock
already exists, attempting to get a lock will block.

Modify the code to ensure that the lock is released. Releasing a lock is
idempotent.

I am not 100% certain this fixes the problem (which is an edge case I do
not know how to reproduce), but the code does still seem to work
otherwise.