SRE-21944: fix error for signalfx with unvalidated deployment notifications
I noticed this error recently, but I don't think this is due to any of
my changes. Given the code, I think this never worked, but since it was
listed last in our config, we didn't notice.

Unvalidated deployment notifications are a reminder, not actual events,
so we don't want them in signalfx anyway.