SRE-21810: ensure that we clear out stale locks
If kazoo raises a ConnectionClosedError, then the overlying TDS code
will have broken out of this loop. We need to re-enter the loop (upon
the next attempt) in order to release any lock which we may or may not
still have.

This should fix a case where multiple daemons end up operating at the
same time after zookeeper connection failures.